refactor: moved TritonServerServiceStepDefinitions to internal package
