+++ kura/test/org.eclipse.kura.ai.triton.server.test/src/test/java/org/eclipse/kura/ai/triton/server/TritonServerServiceEngineTest.java
@@ -15,6 +15,7 @@ package org.eclipse.kura.ai.triton.server;
 
 import java.io.IOException;
 
+import org.eclipse.kura.ai.triton.server.internal.TritonServerServiceStepDefinitions;
 import org.junit.Test;
 
 public class TritonServerServiceEngineTest extends TritonServerServiceStepDefinitions {
