+++ kura/test/org.eclipse.kura.ai.triton.server.test/src/test/java/org/eclipse/kura/ai/triton/server/TritonServerServiceModelTest.java
@@ -16,6 +16,7 @@ package org.eclipse.kura.ai.triton.server;
 import java.io.IOException;
 
 import org.eclipse.kura.KuraException;
+import org.eclipse.kura.ai.triton.server.internal.TritonServerServiceStepDefinitions;
 import org.junit.Test;
 
 public class TritonServerServiceModelTest extends TritonServerServiceStepDefinitions {
